requirements.txt
uwsgi==2.0.15

flask==0.12.2
flask_script==2.0.5
flask_restful==0.3.5
flask_fixtures==0.3.7
flask_sqlalchemy==2.1.0
flask_bcrypt==0.7.1
flask-jwt-extended==2.4.1
simplekv==0.10.0
ldap3

email_validator==1.0.2

sqlalchemy==1.1.10
sqlalchemy_utils==0.32.14
psycopg2==2.7.1

python-slugify==1.2.4

pillow==4.0.0
pytz
babel==2.4.0
